requirements.txt
Django==1.9.5
django-adminactions==1.0
django-controlcenter==0.1.2
django-debug-toolbar==1.4
django-report-builder==3.3.0
django-report-utils==0.3.15
django-suit==0.2.18
djangorestframework==3.3.3
future==0.15.2
jdcal==1.2
openpyxl==2.2.1
psycopg2==2.6.1
python-dateutil==2.5.2
pytz==2016.3
six==1.10.0
sqlparse==0.1.19
unicodecsv==0.14.1
xlrd==0.9.4
xlwt-future==0.8.0
